requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.1.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
msgpack==1.1.0
PyJWT==2.9.0
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.2
